Convolutional Neural Network/requirements.txt
matplotlib
numpy
pandas
Pillow
scikit-image
Opencv-python
globus_sdk
keras
tensorflow
dlhub_sdk